Refactor FindMeetingQuery.java file
Before, the strategy to find free events had a longer runtime and was much more complex. The previous strategy was to divide the day into free options after every single event that involves a mandatory attendee. This commit changes the structure such that it first aggregates all events that contain the mandatory attendees and then merges any conflicts between them.